Refactor: Change BoardStatus,BoardType location, method name to lowercase
